refactor: split step_book.go into 6 files by concern
Split the 2,525-line step_book.go into focused files:
- step_book.go (1,134): struct, constructors, Rev/Fwd, accessors, AMM
- step_book_quality.go (236): quality functions, transfer rates
- step_book_offer.go (476): offer iteration, filtering, funding
- step_book_consume.go (202): offer consumption, deletion
- step_book_transfer.go (192): fund transfers, owner count
- step_book_trustline.go (320): trustline credit/debit/create

Pure file reorganization — zero behavioral changes.

Closes #75
